fix(indexer): address chain-indexer scaffold review findings
- fail bootstrap when the HTTP server cannot start and exit the process
  when the sync loop dies so healthz cannot report healthy on a dead sync
- detect advisory-lock session loss via backend pid checks to prevent
  two sync leaders after a transparent reconnect
- retry transient sync errors with capped backoff; continuity breaks and
  leadership loss stay fatal
- serialize migrations with a pg advisory lock
- select existing message types before inserting to stop id sequence
  burn; widen message_types.id to serial and messages.type_id to integer
- fail fast when txs_results does not match block txs instead of
  persisting txs as successful with zero gas
- serialize Buffers as base64 in canonical json despite Buffer#toJSON
- treat empty SYNC_START_HEIGHT as absent; validate PORT range
- force-close keep-alive sockets on shutdown; handle async listen errors;
  drop the no-op exit listener
- return a constant message for unexpected errors; log via LoggerService
- remove dead InjectTypeRegistry export and inline comments

+++ apps/chain-indexer/src/db/schema.ts
@@ -1,4 +1,4 @@
-import { bigint, index, integer, jsonb, pgSchema, pgTable, primaryKey, smallint, smallserial, text, timestamp, uniqueIndex } from "drizzle-orm/pg-core";
+import { bigint, index, integer, jsonb, pgSchema, pgTable, primaryKey, serial, text, timestamp, uniqueIndex } from "drizzle-orm/pg-core";
 
 import { bytea } from "@src/db/bytea";
 
@@ -35,7 +35,7 @@ export const Transactions = cosmosSchema.table(
 export const MessageTypes = cosmosSchema.table(
   "message_types",
   {
-    id: smallserial("id").primaryKey(),
+    id: serial("id").primaryKey(),
     type: text("type").notNull()
   },
   t => [uniqueIndex("message_types_type_idx").on(t.type)]
@@ -47,7 +47,7 @@ export const Messages = cosmosSchema.table(
     height: bigint("height", { mode: "number" }).notNull(),
     txIndex: integer("tx_index").notNull(),
     index: integer("index").notNull(),
-    typeId: smallint("type_id")
+    typeId: integer("type_id")
       .notNull()
       .references(() => MessageTypes.id),
     body: jsonb("body")
